+++ projects/Companion/Components/ImportSettings.lua
@@ -1,4 +1,4 @@
--- A page for importing palettes
+-- Interface for importing settings
 
 local HttpService = game:GetService("HttpService")
 local RunService = game:GetService("RunService")
@@ -11,18 +11,21 @@ local root = script.Parent.Parent
 local Common = root.Common
 
 local CommonModules = Common.Modules
+local ColorPaneUserDataDefaultValues = require(CommonModules.ColorPaneUserDataDefaultValues)
+local ColorPaneUserDataValidators = require(CommonModules.ColorPaneUserDataValidators)
 local Style = require(CommonModules.Style)
 local Translator = require(CommonModules.Translator)
 
 local CommonIncludes = Common.Includes
+local Cryo = require(CommonIncludes.Cryo)
 local Promise = require(CommonIncludes.Promise)
 local Roact = require(CommonIncludes.RoactRodux.Roact)
-local RoactRodux = require(CommonIncludes.RoactRodux.RoactRodux)
+local t = require(CommonIncludes.t)
 
 local CommonComponents = Common.Components
 local Button = require(CommonComponents.Button)
+local ConnectTheme = require(CommonComponents.ConnectTheme)
 local RadioButtonGroup = require(CommonComponents.RadioButtonGroup)
-local TextInput = require(CommonComponents.TextInput)
 
 local StandardComponents = CommonComponents.StandardComponents
 local StandardScrollingFrame = require(StandardComponents.ScrollingFrame)
@@ -31,28 +34,23 @@ local StandardUIListLayout = require(StandardComponents.UIListLayout)
 local StandardUIPadding = require(StandardComponents.UIPadding)
 
 local Modules = root.Modules
-local Enums = require(Modules.Enums)
-local Util = require(Modules.Util)
+local CompanionUserDataDefaultValues = require(Modules.CompanionUserDataDefaultValues)
+local CompanionUserDataValidators = require(Modules.CompanionUserDataValidators)
+local Constants = require(Modules.Constants)
 
 ---
 
-local HTTP_TIMEOUT = 10
-
-local importTypes = { "ModuleScript", "StringValue", "File", "URL" }
+local importTypes = { "ModuleScript", "StringValue", "File" }
 
 local importOptions = {
     "ModuleScript",
     "StringValue",
     Translator.FormatByKey("JSONFile_ImportType"),
-    Translator.FormatByKey("URL_ImportType"),
 }
 
 local uiTranslations = Translator.GenerateTranslationTable({
     "JSONFile_ImportType",
-    "URL_ImportType",
-    "URL_InputText",
 
-    "RetrieveURL_ButtonText",
     "UseSelection_ButtonText",
     "SelectAFile_ButtonText",
     "Import_ButtonText",
@@ -61,7 +59,21 @@ local uiTranslations = Translator.GenerateTranslationTable({
     "WaitingForImport_Message",
     "EmptySelection_Message",
     "MultipleSelections_Message",
-    "PaletteNameOK_Message",
+})
+
+local settingsTableValidator: (any) -> (boolean, string?) = t.interface({
+    [Constants.COLORPANE_USERDATA_KEY] = t.interface({
+        AskNameBeforePaletteCreation = t.optional(ColorPaneUserDataValidators.AskNameBeforePaletteCreation),
+        SnapValue = t.optional(ColorPaneUserDataValidators.SnapValue),
+        UserColorPalettes = t.optional(ColorPaneUserDataValidators.UserColorPalettes),
+        UserGradientPalettes = t.optional(ColorPaneUserDataValidators.UserGradientPalettes),
+    }),
+
+    [Constants.COMPANION_USERDATA_KEY] = t.interface({
+        AutoLoadColorPropertiesAPIData = t.optional(CompanionUserDataValidators.AutoLoadColorPropertiesAPIData),
+        CacheColorPropertiesAPIData = t.optional(CompanionUserDataValidators.CacheColorPropertiesAPIData),
+        RobloxApiDump = t.optional(CompanionUserDataValidators.RobloxApiDump),
+    }),
 })
 
 local statusIcons = {
@@ -80,41 +92,28 @@ local statusColorGenerators = {
     end,
 }
 
---[[
-local importStatusMessages = {
-    NoObjectSelected = "At least one object should be selected",
-    MultipleObjectsSelected = "Only one object should be selected",
-    NotAnX = "The selected object is not a %s",
-    ValidPalette = "The %s contains a valid palette",
-    NonConformantPalette = "Palette format check failed: %s",
-}
-]]
-
 ---
 
 --[[
     props
-        onPromptClosed: (boolean) -> nil
-    
+
+        onPromptClosed: ({}?) -> ()
+
     store props
+
         theme: StudioTheme
-        palettes: array<Palette>
-
-        addPalette: (Palette) -> nil
 ]]
+local ImportSettings = Roact.PureComponent:extend("ImportSettings")
 
-local ImportPalette = Roact.PureComponent:extend("ImportPalette")
-
-ImportPalette.init = function(self)
+ImportSettings.init = function(self)
     self.statusIcon = Roact.createRef()
     self.listLength, self.updateListLength = Roact.createBinding(0)
 
     self.importSuccessHandler = function(message)
-        return function(palette)
-            if (palette) then
+        return function(settings)
+            if (settings) then
                 self:setState({
-                    palette = palette,
-                    paletteName = palette.name,
+                    settings = settings,
                     
                     status = "ok",
                     statusMessage = message
@@ -140,7 +139,7 @@ ImportPalette.init = function(self)
     end
 end
 
-ImportPalette.didMount = function(self)
+ImportSettings.didMount = function(self)
     self.rotator = RunService.Heartbeat:Connect(function(step)
         local statusIcon = self.statusIcon:getValue()
         if (not statusIcon) then return end
@@ -153,26 +152,17 @@ ImportPalette.didMount = function(self)
     end)
 end
 
-ImportPalette.willUnmount = function(self)
+ImportSettings.willUnmount = function(self)
     self.rotator:Disconnect()
-
-    if (self.webImportPromise) then
-        self.webImportPromise:cancel()
-        self.webImportPromise = nil
-    end
 end
 
-ImportPalette.render = function(self)
+ImportSettings.render = function(self)
     local theme = self.props.theme
 
     local status = self.state.status
     local importType = self.state.importType
+    local settings = self.state.settings
     local importPage
-
-    local palettes = self.props.palettes
-    local palette = self.state.palette
-    local paletteName = self.state.paletteName
-    local newPaletteName = if paletteName then Util.palette.getNewItemName(palettes, paletteName) else nil
 
     if (importType == "ModuleScript") then
         importPage = Roact.createElement("Frame", {
@@ -198,7 +188,7 @@ ImportPalette.render = function(self)
 
                 onActivated = function()
                     self:setState({
-                        palette = Roact.None,
+                        settings = Roact.None,
                     })
 
                     Promise.new(function(resolve, reject)
@@ -219,21 +209,21 @@ ImportPalette.render = function(self)
                             return
                         end
 
-                        local newPalette = require(selection[1])
+                        local newSettingsTable = require(selection[1])
                         
-                        if (type(newPalette) == "string") then
-                            newPalette = HttpService:JSONDecode(newPalette)
+                        if (type(newSettingsTable) == "string") then
+                            newSettingsTable = HttpService:JSONDecode(newSettingsTable)
                         end
 
-                        local isValid, message = Util.palette.validate(newPalette)
+                        local isValid, message = settingsTableValidator(newSettingsTable)
 
                         if (isValid) then
-                            resolve(newPalette)
+                            resolve(newSettingsTable)
                         else
-                            reject(Translator.FormatByKey("NonConformantPalette_Message", { message }))
+                            reject(Translator.FormatByKey("NonConformantSettingsTable_Message", { message }))
                         end
                     end):andThen(
-                        self.importSuccessHandler(Translator.FormatByKey("ValidImport_Message", { "ModuleScript" })),
+                        self.importSuccessHandler(Translator.FormatByKey("ValidSettingsTableImport_Message", { "ModuleScript" })),
                         self.importErrorHandler
                     )
                 end,
@@ -290,7 +280,7 @@ ImportPalette.render = function(self)
 
                 onActivated = function()
                     self:setState({
-                        palette = Roact.None,
+                        settings = Roact.None,
                     })
 
                     Promise.new(function(resolve, reject)
@@ -311,16 +301,16 @@ ImportPalette.render = function(self)
                             return
                         end
 
-                        local newPalette = HttpService:JSONDecode(object.Value)
-                        local isValid, message = Util.palette.validate(newPalette)
+                        local newSettingsTable = HttpService:JSONDecode(object.Value)
+                        local isValid, message = settingsTableValidator(newSettingsTable)
 
                         if (isValid) then
-                            resolve(newPalette)
+                            resolve(newSettingsTable)
                         else
-                            reject(Translator.FormatByKey("NonConformantPalette_Message", { message }))
+                            reject(Translator.FormatByKey("NonConformantSettingsTable_Message", { message }))
                         end
                     end):andThen(
-                        self.importSuccessHandler(Translator.FormatByKey("ValidImport_Message", { "StringValue" })),
+                        self.importSuccessHandler(Translator.FormatByKey("ValidSettingsTableImport_Message", { "StringValue" })),
                         self.importErrorHandler
                     )
                 end,
@@ -371,7 +361,7 @@ ImportPalette.render = function(self)
 
                 onActivated = function()
                     self:setState({
-                        palette = Roact.None,
+                        settings = Roact.None,
                         status = "wait",
                     })
 
@@ -385,16 +375,16 @@ ImportPalette.render = function(self)
                             return
                         end
                         
-                        local newPalette = HttpService:JSONDecode(file:GetBinaryContents())
-                        local isValid, message = Util.palette.validate(newPalette)
+                        local newSettingsTable = HttpService:JSONDecode(file:GetBinaryContents())
+                        local isValid, message = settingsTableValidator(newSettingsTable)
 
                         if (isValid) then
-                            resolve(newPalette)
+                            resolve(newSettingsTable)
                         else
-                            reject(Translator.FormatByKey("NonConformantPalette_Message", { message }))
+                            reject(Translator.FormatByKey("NonConformantSettingsTable_Message", { message }))
                         end
                     end):andThen(
-                        self.importSuccessHandler(Translator.FormatByKey("ValidImport_Message", { "file" })),
+                        self.importSuccessHandler(Translator.FormatByKey("ValidSettingsTableImport_Message", { "file" })),
                         self.importErrorHandler
                     ):finally(function()
                         if (file) then
@@ -432,122 +422,21 @@ ImportPalette.render = function(self)
                 TextWrapped = true,
             }),
         })
-    elseif (importType == "URL") then
-        importPage = Roact.createElement("Frame", {
-            BackgroundTransparency = 1,
-            BorderSizePixel = 0,
-            LayoutOrder = 3,
-            
-            Size = UDim2.new(
-                1, 0, 0, Style.Constants.StandardInputHeight +
-                    Style.Constants.StandardButtonHeight +
-                    Style.Constants.StandardTextSize * 2 +
-                    Style.Constants.MinorElementPadding * 2
-            ),
-        }, {
-            URLInput = Roact.createElement(TextInput, {
-                AnchorPoint = Vector2.new(0, 0),
-                Position = UDim2.new(0, 0, 0, 0),
-                Size = UDim2.new(1, 0, 0, Style.Constants.StandardInputHeight),
-
-                Text = self.state.importURL or "",
-                PlaceholderText = uiTranslations["URL_InputText"],
-
-                onSubmit = function(newText)
-                    self:setState({
-                        importURL = newText
-                    })
-                end,
-
-                canSubmitEmptyString = true,
-                selectTextOnFocus = true,
-            }),
-
-            RetrieveURLButton = Roact.createElement(Button, {
-                AnchorPoint = Vector2.new(0, 0),
-                Position = UDim2.new(0, 0, 0, Style.Constants.StandardInputHeight + Style.Constants.MinorElementPadding),
-                Size = UDim2.new(0, 80, 0, Style.Constants.StandardButtonHeight),
-                
-                disabled = (status == "wait"),
-                displayType = "text",
-                text = uiTranslations["RetrieveURL_ButtonText"],
-
-                onActivated = function()
-                    local url = self.state.importURL
-                    if (not url) or (url == "") then return end
-
-                    if (self.webImportPromise) then
-                        self.webImportPromise:cancel()
-                        self.webImportPromise = nil
-                    end
-
-                    self:setState({
-                        palette = Roact.None,
-                        status = "wait"
-                    })
-
-                    local promise = Promise.new(function(resolve, reject)
-                        local response = HttpService:RequestAsync({
-                            Url = url,
-                            Method = "GET"
-                        })
-
-                        if (not response.Success) then
-                            reject(response.StatusCode .. " " .. response.StatusMessage)
-                            return
-                        else
-                            local newPalette = HttpService:JSONDecode(response.Body)
-                            local isValid, message = Util.palette.validate(newPalette)
-
-                            if (isValid) then
-                                resolve(newPalette)
-                            else
-                                reject(Translator.FormatByKey("NonConformantPalette_Message", { message }))
-                            end
-                        end
-                    end):timeout(HTTP_TIMEOUT)
-                    
-                    promise:andThen(
-                        self.importSuccessHandler(Translator.FormatByKey("ValidImport_Message", { "URL" })),
-                        self.importErrorHandler
-                    ):finally(function()
-                        self.webImportPromise = nil
-                    end)
-
-                    self.webImportPromise = promise
-                end,
-            }),
-
-            StatusIcon = Roact.createElement("ImageLabel", {
-                AnchorPoint = Vector2.new(0, 0),
-                Size = Style.UDim2.StandardButtonSize,
-                Position = UDim2.new(0, 80 + Style.Constants.MinorElementPadding, 0, Style.Constants.StandardInputHeight + Style.Constants.MinorElementPadding),
-                BackgroundTransparency = 1,
-                BorderSizePixel = 0,
-
-                Image = if status then statusIcons[status] else "",
-
-                ImageColor3 = if statusColorGenerators[status] then
-                    statusColorGenerators[status](theme)
-                else theme:GetColor(Enum.StudioStyleGuideColor.MainText),
-
-                [Roact.Ref] = self.statusIcon,
-            }),
-
-            StatusText = Roact.createElement(StandardTextLabel, {
-                AnchorPoint = Vector2.new(0, 0),
-                Size = UDim2.new(1, 0, 0, Style.Constants.StandardTextSize * 2),
-                Position = UDim2.new(0, 0, 0, Style.Constants.StandardInputHeight + Style.Constants.StandardButtonHeight + (Style.Constants.MinorElementPadding * 2)),
-
-                Text = self.state.statusMessage or uiTranslations["WaitingForImport_Message"],
-                TextXAlignment = Enum.TextXAlignment.Left,
-                TextYAlignment = Enum.TextYAlignment.Top,
-                TextWrapped = true,
-            }),
-        })
     end
 
-    return Roact.createFragment({
+    return Roact.createElement("Frame", {
+        AnchorPoint = Vector2.new(0.5, 0.5),
+        Position = UDim2.new(0.5, 0, 0.5, 0),
+        Size = UDim2.new(1, 0, 1, 0),
+        BackgroundTransparency = 0,
+        BorderSizePixel = 0,
+
+        BackgroundColor3 = theme:GetColor(Enum.StudioStyleGuideColor.MainBackground),
+    }, {
+        UIPadding = Roact.createElement(StandardUIPadding, {
+            paddings = {Style.Constants.PagePadding}
+        }),
+
         Dialog = Roact.createElement(StandardScrollingFrame, {
             AnchorPoint = Vector2.new(0.5, 0),
             Size = UDim2.new(1, 0, 1, -(Style.Constants.StandardButtonHeight + Style.Constants.SpaciousElementPadding)),
@@ -574,7 +463,7 @@ ImportPalette.render = function(self)
             }),
             
             ImportType = Roact.createElement(RadioButtonGroup, {
-                Size = UDim2.new(1, 0, 0, (Style.Constants.StandardButtonHeight * 4) + (Style.Constants.MinorElementPadding * 3)),
+                Size = UDim2.new(1, 0, 0, (Style.Constants.StandardButtonHeight * 3) + (Style.Constants.MinorElementPadding * 2)),
                 LayoutOrder = 1,
     
                 selected = table.find(importTypes, importType),
@@ -583,22 +472,15 @@ ImportPalette.render = function(self)
                 onSelected = function(i)
                     self:setState({
                         importType = importTypes[i],
-
-                        importURL = Roact.None,
-                        palette = Roact.None,
+                        settings = Roact.None,
 
                         status = Roact.None,
                         statusMessage = Roact.None,
                     })
-
-                    if (self.webImportPromise) then
-                        self.webImportPromise:cancel()
-                        self.webImportPromise = nil
-                    end
                 end,
             }),
 
-            Separator1 = if self.state.importType then
+            Separator1 = if importType then
                 Roact.createElement("Frame", {
                     Size = UDim2.new(1, 0, 0, 1),
                     BackgroundTransparency = 0,
@@ -609,52 +491,7 @@ ImportPalette.render = function(self)
                 })
             else nil,
 
-            ImportPage = if self.state.importType then importPage else nil,
-
-            Separator2 = if self.state.palette then
-                Roact.createElement("Frame", {
-                    Size = UDim2.new(1, 0, 0, 1),
-                    BackgroundTransparency = 0,
-                    BorderSizePixel = 0,
-                    LayoutOrder = 4,
-
-                    BackgroundColor3 = theme:GetColor(Enum.StudioStyleGuideColor.Separator)
-                })
-            else nil,
-
-            Naming = if self.state.palette then
-                Roact.createElement("Frame", {
-                    Size = UDim2.new(1, 0, 0, Style.Constants.StandardInputHeight + Style.Constants.StandardTextSize + Style.Constants.MinorElementPadding),
-                    BackgroundTransparency = 1,
-                    BorderSizePixel = 0,
-                    LayoutOrder = 5,
-                }, {
-                    NameInput = Roact.createElement(TextInput, {
-                        AnchorPoint = Vector2.new(0, 0),
-                        Position = UDim2.new(0, 0, 0, 0),
-                        Size = UDim2.new(1, 0, 0, Style.Constants.StandardInputHeight),
-            
-                        Text = self.state.paletteName,
-                        TextSize = Style.Constants.StandardTextSize,
-
-                        onSubmit = function(newText)
-                            self:setState({
-                                paletteName = newText
-                            })
-                        end,
-                    }),
-            
-                    NameIsOKLabel = Roact.createElement(StandardTextLabel, {
-                        AnchorPoint = Vector2.new(0, 0),
-                        Position = UDim2.new(0, 0, 0, Style.Constants.StandardInputHeight + Style.Constants.MinorElementPadding),
-                        Size = UDim2.new(1, 0, 0, Style.Constants.StandardTextSize),
-
-                        Text = if (paletteName ~= newPaletteName) then
-                            Translator.FormatByKey("PaletteRename_Message", { newPaletteName })
-                        else uiTranslations["PaletteNameOK_Message"],
-                    }),
-                })
-            else nil,
+            ImportPage = if importType then importPage else nil,
         }),
 
         Buttons = Roact.createElement("Frame", {
@@ -683,7 +520,7 @@ ImportPalette.render = function(self)
                 displayColor = theme:GetColor(Enum.StudioStyleGuideColor.DialogButtonText),
 
                 onActivated = function()
-                    self.props.onPromptClosed(false)
+                    self.props.onPromptClosed(nil)
                 end
             }),
 
@@ -691,7 +528,7 @@ ImportPalette.render = function(self)
                 Size = Style.UDim2.DialogButtonSize,
                 LayoutOrder = 1,
 
-                disabled = (not (palette and newPaletteName)),
+                disabled = (not settings),
                 displayType = "text",
                 text = uiTranslations["Import_ButtonText"],
 
@@ -701,38 +538,16 @@ ImportPalette.render = function(self)
                 displayColor = theme:GetColor(Enum.StudioStyleGuideColor.DialogMainButtonText),
 
                 onActivated = function()
-                    palette = Util.table.deepCopy(palette)
-                    palette.name = newPaletteName
+                    local filledSettingsTable = {
+                        [Constants.COLORPANE_USERDATA_KEY] = Cryo.Dictionary.join(ColorPaneUserDataDefaultValues, settings[Constants.COLORPANE_USERDATA_KEY]),
+                        [Constants.COMPANION_USERDATA_KEY] = Cryo.Dictionary.join(CompanionUserDataDefaultValues, settings[Constants.COMPANION_USERDATA_KEY]),
+                    }
 
-                    -- convert tables to colors
-                    local colors = palette.colors
-
-                    for i = 1, #colors do
-                        local color = colors[i].color
-                        
-                        colors[i].color = Color3.new(color[1], color[2], color[3])
-                    end
-
-                    self.props.addPalette(palette)
-                    self.props.onPromptClosed(true)
+                    self.props.onPromptClosed(filledSettingsTable)
                 end
             }),
         })
     })
 end
 
-return RoactRodux.connect(function(state)
-    return {
-        theme = state.theme,
-        palettes = state.colorEditor.palettes,
-    }
-end, function(dispatch)
-    return {
-        addPalette = function(palette)
-            dispatch({
-                type = Enums.StoreActionType.ColorEditor_AddPalette,
-                palette = palette,
-            })
-        end,
-    }
-end)(ImportPalette)
+return ConnectTheme(ImportSettings)